support renaming and fix logic in rename_interfaces
The one issue i'm aware of currently is that tap devices
 (ip tuntap add mode tap user root mytap1)
do not work correctly with 'is_up' which means the check
does not bring them down and the rename fails.

The LOG.debug message should be cleaned up too, as it currently
references the function rather function.__name__ for nicer message.